feature: 자격증 업데이트 구현

+++ src/main/java/com/zigtong/clientserver/domain/certificate/repository/CertificateRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.zigtong.clientserver.domain.certificate.entity.Certificate;
 
-public interface CertificateRepository extends JpaRepository<Certificate, Long> {
+public interface CertificateRepository extends JpaRepository<Certificate, Integer> {
 	List<Certificate> findAllByAffiliation(String affiliationName);
 
 	@Query("SELECT DISTINCT c.affiliation FROM Certificate c")
